Fixed delete user issue. Problem when follow yourself

+++ app.py
@@ -26,8 +26,6 @@ app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY', "it's a secret")
 toolbar = DebugToolbarExtension(app)
 
 connect_db(app)
-# db.create_all()
-
 
 ##############################################################################
 # User signup/login/logout
@@ -41,6 +39,7 @@ def add_user_to_g():
         g.user = User.query.get(session[CURR_USER_KEY])
         g.logout_form = LogoutForm()
         g.like_form = LikeMessageForm()
+        g.delete_user_form = DeleteUserForm()
 
     else:
         g.user = None
@@ -264,8 +263,9 @@ def delete_user():
 
     do_logout()
 
-    db.session.delete(g.user)
-    db.session.commit()
+    if g.delete_user_form.validate_on_submit():
+        db.session.delete(g.user)
+        db.session.commit()
 
     return redirect("/signup")
